refactor: simplify repl cli
